+++ assets/javascripts/kitten/components/cards/contribution-card/components/action.js
@@ -1,29 +1,30 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import { Button } from '../../../../components/buttons/button'
 import { CheckedCircleIcon } from '../../../icons/checked-circle-icon'
 import COLORS from '../../../../constants/colors-config'
+import { Context } from '../context'
 
-export const Action = ({ isInputValid, title, ...props }) => {
+export const Action = ({ isInputValid, title, className, ...props }) => {
+  const { isInputEmpty } = useContext(Context)
 
   return (
     <>
-      <Button
-        borderRadius={4}
-        type="submit"
-        {...props}
-        className={classNames(
-          'k-ContributionCard__action',
-          { 'k-ContributionCard__actionHide': isInputValid },
-          props.className,
-        )}
-      />
-      {isInputValid && (
+      {isInputValid ? (
         <CheckedCircleIcon
           bgColor={COLORS.valid}
           color={COLORS.background1}
           title={title}
+        />
+      ) : (
+        <Button
+          borderRadius={4}
+          type="submit"
+          className={classNames('k-ContributionCard__action', className)}
+          modifier="helium"
+          disabled={isInputEmpty}
+          {...props}
         />
       )}
     </>
@@ -32,7 +33,7 @@ export const Action = ({ isInputValid, title, ...props }) => {
 
 Action.defaultProps = {
   isInputValid: false,
-  title: "",
+  title: '',
 }
 
 Action.propTypes = {