+++ assets/javascripts/kitten/components/cards/contribution-card/components/pill-number.js
@@ -6,10 +6,7 @@ export const PillNumber = props => {
   return (
     <PillNumberInput
       {...props}
-      className={classNames(
-        'k-ContributionCard__pillNumber',
-        props.className,
-      )}
+      className={classNames('k-ContributionCard__pillNumber', props.className)}
     />
   )
 }